refactors cold migration / resize
Simplifies the cold resize / migration process by making the
destination node pull the VM files from the source instead of
vice-versa. Doing so, the logic is simplified greatly, the
destination node only has to copy the files, regardless if it
is on shared storage or not.

Solves the migration issue when the compute nodes have different
instances_path config option set, as the destination node will
copy the files to its configured instances_path folder.

Co-Authored-By: Claudiu Belu <cbelu@cloudbasesolutions.com>

Change-Id: Ia44850675e8cb8eb21017e1c4321802574dfe6a0

refactors vmops' create_instance

Splits the create_instance method into two, separating the
"updateable" resources from the method. This new method will be
called from migrationops in the following patch, when performing a resize.

Change-Id: I94277e8f1a58ecc1529f3cf92deac3596f5b65d9

adds resize option to updatable resources

When importing a VM during a cold resize, the VM's
resources needs to be updated as well. This patch will
facilitate the cold migration refactor.

Depends-On: Ia7fbb2a9bc8f5d21bf5ec4a865fc7b1368ecf136
Depends-On: Ic153dcfd621a1e6235dd3a75241274ee7d7c868b
Depends-On: Ic9248f44de6ff097b8080a80fd3212eeef6712f6
Depends-On: Ib83bbc5c7fd07b55888f58d3407e1dd1a50e9b7e
Depends-On: Iec9b501880098e69be652bf6c2a90a9864e52a9c

Hyper-V: Resize refactoring

Resizing Shielded VMs wasn't possible since the vm was destroyed
and recreated. This patch modifies the way resize is done in order
to add support for resizing Shielded VMs as well.

Co-Authored-By: Claudiu Belu <cbelu@cloudbasesolutions.com>

Depends-On: I01efbc312d8ea81c3a4e1fd82cccc695c7472b5f

Partially-Implements: blueprint hyper-v-shielded-vms

Change-Id: Ia33031395ee25e348339f1e756ef395d3f63e33c

base instance migration revert dir on the instance path

get_instance_migr_revert_dir bases the returned path on the
configured instaces_path. In the next commit, the instances won't
necessarely exist on a host, or in the configured instances_path.

Change-Id: Ifeaef0569e216943fb7766cc681293f4958cb37f

don't move files on migration

In some cases (shared storage, CSVs), moving the instance
files on cold migration / resize might not be desired.

Adds move_disks_on_cold_migration config option (default value True).

If the config option is True, or if the source instance path is not
on a CSV or a share, the instance files are copied to the destination
node. Otherwise, the files are kept on the source node.

Change-Id: Ia093d6e68d033d060afff863c22ef2d97490b0ff

Conflicts:
	hyperv/nova/migrationops.py
	hyperv/nova/pathutils.py
	hyperv/nova/vmops.py
	hyperv/nova/volumeops.py
	hyperv/tests/unit/test_migrationops.py
	hyperv/tests/unit/test_vmops.py